Optimize branch misses in AssociativeApplierImpl::apply
The short-circuit evaluation was implemented when applying the
saturable operators (and, or) on a vector of ColumnUInt8. However,
its control flow would be compiled as a series of conditional
branch instructions which are hard to predict by the hardware and
as a result, the overall performance is lowered. To remedy this
issue, the short-circuit is removed and the whole expression is
evaluated instead.

A microbenchmark is implemented to evaluate the performance of the
AssociativeApplierImpl::apply method.