Implement TreeDNFConverter
This commit implements the TreeDNFConverter which transforms the
logic expressions to their DNF (Disjunctive Normal Form) for the
efficient optimization of the mergeable predicates concanated by
AND operations.

As the converters of CNF/DNF share a lot of basic operations on
the data structures, the DNF converter is implemented in the same
files as the CNF one.